Add barrel index.ts files
