+++ src/App.tsx
@@ -1,5 +1,5 @@
-import WorkspacePanel from './ui/workspace/WorkspacePanel'
-import EditorModal from "./ui/editor/modal/EditorModal.tsx"
+import {WorkspacePanel} from "./ui/workspace"
+import {EditorModal} from "./ui/editor"
 import {Box, CssBaseline, ThemeProvider} from "@mui/material";
 import theme from "./theme.ts";
