--- src/core/blockly/blockColours.ts
@@ -1,11 +1,0 @@
-export const colours: Record<string, string> = {
-  events: '#ffbf00',
-  control: '#ffab19',
-  commands: '#4c97ff',
-  literals: '#006676',
-  constructs: '#3aa57a',
-  variable: '#ff8c1a',
-  procedures: '#ff6680',
-  execute: '#4c97ff',
-  targetSelectors: '#5cb1d6',
-}
