+++ src/core/blockly/setupToolbox.ts
@@ -1,0 +1,7 @@
+import {registerBlockSpecs} from "./specs/registry.ts";
+import {applyCategoryColours} from "./colours.ts";
+
+export function setupToolbox() {
+  applyCategoryColours()
+  registerBlockSpecs()
+}
