+++ src/core/blockly/specs/categories/selectors.ts
@@ -1,9 +1,9 @@
 import * as Blockly from 'blockly'
-import { colours } from '../blockColours'
-import { ToggleImageField } from '../fields/toggleImage'
-import { mcfunctionGenerator } from '../../compiler/blockly/generator'
-import type {BlockShadowStatesFunction, BlockSpec} from './types'
-import { setShadowState } from '../extensions/shadows.ts'
+import { colours } from '../../colours'
+import { ToggleImageField } from '../../fields/toggleImage'
+import { mcfunctionGenerator } from '../../../compiler'
+import type {BlockShadowStatesFunction, BlockSpec} from '../types'
+import { setShadowState } from '../../extensions/shadows.ts'
 
 const INPUT_FILTER_STACK = 'FILTER_STACK'
 const FIELD_BASE = 'BASE'
