+++ src/core/catalog/index.ts
@@ -1,0 +1,2 @@
+export * from './itemCatalog'
+export * from './dataComponentSchemaCatalog'
