+++ src/core/compiler/snbt.ts
