+++ src/core/editor/index.ts
@@ -1,0 +1,2 @@
+export * from './testEditors.ts'
+export * from './types.ts'
