+++ src/core/editor/testEditors.ts
@@ -1,7 +1,6 @@
 import type {AnyEditorState, AnyEditorStateCallback, EditorSchema} from "./types.ts"
 import {useEffect, useState} from "react";
-import {controller} from "../../ui/editor/modal/controller.ts";
-import loadFromSchema from "../../ui/editor/loadFromSchema.tsx";
+import {controller, loadFromSchema} from "../../ui/editor";
 import compileEditorState, {inferCompilerType} from "../compiler/editor/compileEditorState.ts";
 
 export function useTestString() {
