+++ src/ui/editor/index.ts
@@ -1,0 +1,3 @@
+export {default as loadFromSchema} from './loadFromSchema.tsx'
+export {default as EditorModal} from './modal/EditorModal.tsx'
+export {controller} from './modal/controller.ts'
