+++ src/ui/workspace/WorkspacePanel.tsx
@@ -1,73 +1,15 @@
-import { useEffect, useRef } from 'react'
-import * as Blockly from 'blockly'
-import JSZip from 'jszip'
-import {registerContinuousToolbox, additionalOptions, setupWorkspace} from '../../core/blockly/workspaceSetup.ts'
-import getToolboxContents from '../../core/blockly'
-import { compile } from '../../core/compiler'
-import { updateWorkspaceRegistry} from '../../core/compiler/workspaceRegistry'
 import './WorkspacePanel.css'
-
-import type {WorkspaceSvg} from "blockly";
-import {useProjectConfigStore} from "../../stores/projectConfig.ts";
+import JSZip from 'jszip'
+import { compile } from '../../core/compiler'
+import {useProjectConfigStore} from "../../stores";
 import {Box, Button, Checkbox, FormControlLabel, Stack} from "@mui/material"
+import useBlocklyWorkspace from "./useBlocklyWorkspace.ts";
 
-// Register outside useEffect to avoid error due to variables category
-registerContinuousToolbox()
-
-function updateToolbox(workspace: WorkspaceSvg) {
-  updateWorkspaceRegistry(workspace)
-  workspace.updateToolbox({
-    kind: 'categoryToolbox',
-    contents: getToolboxContents(workspace)
-  })
-}
 
 function WorkspacePanel() {
-  const divRef = useRef<HTMLDivElement>(null)
-  const workspaceRef = useRef<Blockly.WorkspaceSvg | null>(null)
-
+  const { divRef, workspaceRef } = useBlocklyWorkspace()
   const projectConfig = useProjectConfigStore((state) => state.projectConfig)
   const updateConfig = useProjectConfigStore((state) => state.updateConfig)
-
-  useEffect(() => {
-    if (!divRef.current) return
-
-    workspaceRef.current = Blockly.inject(divRef.current, {
-      toolbox: {
-        kind: 'categoryToolbox',
-        contents: getToolboxContents()
-      },
-      ...additionalOptions
-    })
-    setupWorkspace(workspaceRef.current!)
-
-    // Initial placeholder variable
-    workspaceRef.current.getVariableMap().createVariable('myVar', 'mc_scoreboard_variable')
-    updateToolbox(workspaceRef.current)
-
-    workspaceRef.current.registerButtonCallback('CREATE_VARIABLE', () => {
-      Blockly.Variables.createVariableButtonHandler(workspaceRef.current!, undefined, 'mc_scoreboard_variable')
-    })
-
-    workspaceRef.current.addChangeListener((event) => {
-      if (
-        // Update toolbox and variable list whenever variables are modified
-        event.type === Blockly.Events.VAR_CREATE ||
-        event.type === Blockly.Events.VAR_DELETE ||
-        event.type === Blockly.Events.VAR_RENAME ||
-        // Update procedure list whenever procedures are modified
-        event.type === Blockly.Events.BLOCK_CREATE ||
-        event.type === Blockly.Events.BLOCK_DELETE ||
-        event.type === Blockly.Events.BLOCK_CHANGE
-      ) {
-        updateToolbox(workspaceRef.current!)
-      }
-    })
-
-    return () => {
-      workspaceRef.current?.dispose()
-    }
-  }, [])
 
   // Debug
   function handleInspect() {
