+++ src/ui/workspace/index.ts
@@ -1,0 +1,1 @@
+export {default as WorkspacePanel} from './WorkspacePanel.tsx'
